[CHANGE]  * Merged the two Fiber dataset idioms: One fast for displaying fibers named now WDataSetFibers and one slow for clustering et. al. The latter is now called WDataSetFiberVector.  * Merged also the two fiber loaders into one WReaderFiberVTK generating a WDataSetFibers  * Merged the two Display modules: Basically the old WMFiberDisplay2 is now the WMFiberDisplay module. The other display module seems to be obsolete, but be aware the the new module has some issues with thread safety and responsiveness.
Note there has to be done the conversion from a WDataSetFibers to a WDataSetFiberVector. This is not commited yet, but will be soon.

+++ src/modules/fiberDisplay/WRMROIRepresentation.cpp
@@ -152,7 +152,7 @@ void WRMROIRepresentation::boxTest( int left, int right, int axis )
     }
 }
 
-unsigned int WRMROIRepresentation::getLineForPoint( unsigned int point )
+size_t WRMROIRepresentation::getLineForPoint( size_t point )
 {
     return m_currentReverse->at( point );
 }
